New class: Workshift
The Workshift class represents a period of time spent working.
The object has methods for start and stop.
It keeps track of the duration it was active, and returns these values
as a dictionary with hours and minutes on display.

Since the user may be working on things in parallel, it's possible
to add a project name to the object for later making sense of the
workshift.

New class: Logbook

The Logbook class is meant as the library for Workshift objects.
Quick lookups, additions and removal of workshifts.

Feature:

Added methods for callback bindings